backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
mysqlclient
python-dotenv
requests
pytz
gunicorn
django-cors-headers